pyproject.toml
[project]
name = "multiagent-system"
version = "0.1.0"
description = "A multi-agent system using Strands Agents framework with word guessing game"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "asyncio-mqtt>=0.11.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
multiagent = "multiagent_system.main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
